consolidate resource authz ops
