+++ orchestrator/careplanservice/authz_careplan.go
@@ -1,8 +1,6 @@
 package careplanservice
 
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
+import "github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
 
 func ReadCarePlanAuthzPolicy() Policy[fhir.CarePlan] {
 	return CareTeamMemberPolicy[fhir.CarePlan]{}
