--- orchestrator/careplanservice/handle_createcondition.go
@@ -1,12 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/SanteonNL/orca/orchestrator/cmd/profile"
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func CreateConditionAuthzPolicy(profile profile.Provider) Policy[fhir.Condition] {
-	return LocalOrganizationPolicy[fhir.Condition]{
-		profile: profile,
-	}
-}
