--- orchestrator/careplanservice/handle_createquestionnaire.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func CreateQuestionnaireAuthzPolicy() Policy[fhir.Questionnaire] {
-	return EveryoneHasAccessPolicy[fhir.Questionnaire]{}
-}
