--- orchestrator/careplanservice/handle_createquestionnaireresponse.go
@@ -1,12 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/SanteonNL/orca/orchestrator/cmd/profile"
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func CreateQuestionnaireResponseAuthzPolicy(profile profile.Provider) Policy[fhir.QuestionnaireResponse] {
-	return LocalOrganizationPolicy[fhir.QuestionnaireResponse]{
-		profile: profile,
-	}
-}
