--- orchestrator/careplanservice/handle_getquestionnaire.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func ReadQuestionnaireAuthzPolicy() Policy[fhir.Questionnaire] {
-	return EveryoneHasAccessPolicy[fhir.Questionnaire]{}
-}
