--- orchestrator/careplanservice/handle_updatepatient.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func UpdatePatientAuthzPolicy() Policy[fhir.Patient] {
-	return CreatorPolicy[fhir.Patient]{}
-}
