--- orchestrator/careplanservice/handle_updatequestionnaire.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func UpdateQuestionnaireAuthzPolicy() Policy[fhir.Questionnaire] {
-	return EveryoneHasAccessPolicy[fhir.Questionnaire]{}
-}
