--- orchestrator/careplanservice/handle_updatequestionnaireresponse.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func UpdateQuestionnaireResponseAuthzPolicy() Policy[fhir.QuestionnaireResponse] {
-	return CreatorPolicy[fhir.QuestionnaireResponse]{}
-}
