--- orchestrator/careplanservice/handle_updateservicerequest.go
@@ -1,9 +1,0 @@
-package careplanservice
-
-import (
-	"github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir"
-)
-
-func UpdateServiceRequestAuthzPolicy() Policy[fhir.ServiceRequest] {
-	return CreatorPolicy[fhir.ServiceRequest]{}
-}
